GitHub API now requires a User-Agent string, so added it via GithubProvider.makeApiRequest.
For example: GithubProvider.makeApiRequest("user", "access_token" -> accessToken)

I re-worked quite a few lines to make it easier (for me) to debug and track down this issue, but the real change is just to ensure the User-Agent is sent.

I also added an Accept header to set the API version to be v3, which GitHub are encouraging.